Petar merge into main

Approved-by: Claude Cournoyer-Cloutier

+++ src/flash/source/Particles/ParticlesMain/active/Sink/Couple_AMUSE/windGinj/Sink/Particles_sinkData.F90
@@ -50,7 +50,7 @@ module Particles_sinkData
   real, save, allocatable, dimension(:,:), target :: particles_local
   real, save, allocatable, dimension(:,:), target :: particles_global
   
-  integer, save :: local_tag_number
+  integer*8, save :: local_tag_number
   integer, save, target :: localnp, localnpf
 
   ! variables needed for Ewald correction in case of periodic BCs
@@ -66,7 +66,7 @@ module Particles_sinkData
   ! Flash evolution step. - JW
   
   integer, save, target :: number_new_sinks
-  integer, save, allocatable, dimension(:), target :: new_sink_tags
+  integer*8, save, allocatable, dimension(:), target :: new_sink_tags
   logical, save :: new_evolution_step = .true.
   real, save :: min_wind_dt = 1e99
   real, save :: wind_inj_rad ! Wind injection radius.
